% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.6 e4.R
\name{AfisareMomente}
\alias{AfisareMomente}
\title{Afișează primele 4 momente inițiale și primele 4 momente centrate ale unei funcții de densitate de probabilitate continue.}
\usage{
AfisareMomente(f)
}
\arguments{
\item{f}{funcția de densitate de probabilitate}
}
\description{
Afișează primele 4 momente inițiale și primele 4 momente centrate ale unei funcții de densitate de probabilitate continue.
}
\examples{
f1 <- function(x) {
  if (x > 0 && x < 20)
    return(1/20)
  else
    return(0)
}
AfisareMomente(f1)
}
