% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.4 e5.R
\name{calcul_medie}
\alias{calcul_medie}
\title{Calculează media variabilei aleatoare g(X)}
\usage{
calcul_medie(g, f, lower = -Inf, upper = Inf)
}
\arguments{
\item{g}{Funcția g specificată de utilizator}

\item{f}{Funcția de densitate de probabilitate a variabilei X}

\item{lower}{Limita inferioară pentru calcul (implicit: -Inf)}

\item{upper}{Limita superioară pentru calcul (implicit: Inf)}
}
\value{
Media variabilei aleatoare g(X)
}
\description{
Calculează media variabilei aleatoare g(X)
}
