% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.4 e5.R
\name{calcul_medie_dispersie}
\alias{calcul_medie_dispersie}
\title{Calculează media și dispersia unei variabile aleatoare g(X), unde X are o repartiție continuă cunoscută și g este o funcție continuă specificată de utilizator.}
\usage{
calcul_medie_dispersie(g, f, lower = -Inf, upper = Inf)
}
\arguments{
\item{g}{Funcția g specificată de utilizator}

\item{f}{Funcția de densitate de probabilitate a variabilei X}

\item{lower}{Limita inferioară pentru calcul (implicit: -Inf)}

\item{upper}{Limita superioară pentru calcul (implicit: Inf)}
}
\value{
Un obiect de tip listă care conține valorile mediei și dispersiei
}
\description{
Calculează media și dispersia unei variabile aleatoare g(X), unde X are o repartiție continuă cunoscută și g este o funcție continuă specificată de utilizator.
}
\examples{
# Funcția de densitate de probabilitate a variabilei X (de exemplu, distribuția normală)
f <- function(x) {
  dnorm(x, mean = 0, sd = 1)
}

# Funcția g specificată de utilizator (de exemplu, g(x) = x^2)
g <- function(x) {
  x^2
}

# Calculul mediei și dispersiei variabilei aleatoare g(X)
rezultate <- calcul_medie_dispersie(g, f)
print(rezultate$media)
print(rezultate$dispersie)
}
