% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.7 e6.R
\name{cRV}
\alias{cRV}
\title{Continuous Random Variable}
\usage{
cRV(pdf)
}
\arguments{
\item{pdf}{probability density function}
}
\value{
cRV object
}
\description{
Continuous Random Variable
}
\examples{
pdf1 <- function(x) { (0 <= x & x <= 1) * (3 * x^2) }
X <- cRV(pdf1)
}
