% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.6 e4.R
\name{dispersia}
\alias{dispersia}
\title{Calculează dispersia unei funcții de densitate de probabilitate continue.}
\usage{
dispersia(f)
}
\arguments{
\item{f}{funcția de densitate de probabilitate}
}
\value{
valoarea dispersiei
}
\description{
Calculează dispersia unei funcții de densitate de probabilitate continue.
}
\examples{
dispersia(function(x) { x^2 })
}
