% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.5 e8.R
\name{distributie_cond}
\alias{distributie_cond}
\title{Distribuție condiționată dintr-un pdf bivariat}
\usage{
distributie_cond(pdf, x = NULL, y = NULL)
}
\arguments{
\item{pdf}{trebuie să fie o funcție de densitate de probabilitate pentru o variabilă aleatoare bivariată}

\item{x}{valoarea fixă a lui X. Furnizarea x va avea ca rezultat f(Y | X = x)}

\item{y}{valoarea fixă a lui Y. Furnizarea x va avea ca rezultat f(X | Y = x)}
}
\description{
Distribuție condiționată dintr-un pdf bivariat
}
\examples{
pdf2 <- function(x, y) { (0 <= x & x <= 1 & 0 <= y & y <= 1) * (2/3 * (x + 2*y)) }

# Distribuția condiționată a lui X dat fiind Y
f_XY <- distributie_cond(pdf2, y = 0.3)
f_XY(0.6)

# Distribuția condiționată a lui Y dat fiind X
f_YX <- distributie_cond(pdf2, x = 0.1)
f_YX(0.75)
}
