% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2 e2.R
\name{gaseste_constanta_normalizare}
\alias{gaseste_constanta_normalizare}
\title{Returnează constanta de normalizare k pentru o funcție dacă aceasta există, în caz contrar returnează null.}
\usage{
gaseste_constanta_normalizare(func)
}
\arguments{
\item{func}{Trebuie să fie o funcție pentru care dorim să găsim constanta de normalizare}
}
\value{
Valoarea integralei (constanta de normalizare k)
}
\description{
Returnează constanta de normalizare k pentru o funcție dacă aceasta există, în caz contrar returnează null.
}
\examples{
EXEMPLU
# Definirea funcției pentru care dorim să găsim constanta de normalizare
f <- function(x) {
  exp((-x^2) / 2)
}

# Apelarea funcției pentru a găsi constanta de normalizare
constant_de_normalizare <- gaseste_constanta_normalizare(f)

# Verificarea rezultatului
if (!is.null(constant_de_normalizare)) {
  cat("Constanta de normalizare k:", constant_de_normalizare, "\n")
} else {
  cat("Nu a fost găsită o constantă de normalizare.\n")
}
}
