% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.7 e6.R
\name{>.cRV}
\alias{>.cRV}
\title{Greater Than Operator for cRV Objects}
\usage{
\method{>}{cRV}(X, x)
}
\arguments{
\item{X}{cRV object}

\item{x}{numeric value}
}
\value{
cRVrez object
}
\description{
Greater Than Operator for cRV Objects
}
