% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.6 e4.R
\name{Initial}
\alias{Initial}
\title{Calculează momentul inițial de ordinul specificat al unei funcții de densitate de probabilitate continue.}
\usage{
Initial(f, ordin)
}
\arguments{
\item{f}{funcția de densitate de probabilitate}

\item{ordin}{ordinul momentului inițial}
}
\value{
valoarea momentului inițial
}
\description{
Calculează momentul inițial de ordinul specificat al unei funcții de densitate de probabilitate continue.
}
\examples{
Initial(function(x) { x^2 }, 4)
}
