% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1 e1.R
\name{is_fdp}
\alias{is_fdp}
\title{Returnează adevărat dacă funcția furnizată este o funcție de densitate de probabilitate și fals în caz contrar}
\usage{
is_fdp(func)
}
\arguments{
\item{is_fpd}{Este funcția pe care vrem să o analizăm}
}
\value{
O valoare booleană care reprezintă dacă funcția furnizată este o funcție de densitate de probabilitate
}
\description{
Returnează adevărat dacă funcția furnizată este o funcție de densitate de probabilitate și fals în caz contrar
}
\examples{
 fmp_valida <- function(x) {
  probabilitati <- c(0.5, 0.3, 0.2, rep(0, 97))  # Probabilitățile pentru valorile 1, 2, 3

  ifelse(x \%in\% 1:3, probabilitati[x], 0)
}

rezultat_fmp <- is_fdp(fmp_valida)
print(rezultat_fmp)

#INVALIDA
fmp_invalida <- function(x) {
  ifelse(x \%\% 2 == 0, 0.5, 0.3)
}

rezultat_fmp <- is_fdp(fmp_invalida)
print(rezultat_fmp)
}
