% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.5 e8.R
\name{is_joint_pdf}
\alias{is_joint_pdf}
\title{Verifică dacă o funcție reprezintă un pdf bivariat}
\usage{
is_joint_pdf(pdf)
}
\arguments{
\item{pdf}{funcția de verificat}
}
\value{
TRUE dacă funcția este un pdf bivariat, FALSE în caz contrar
}
\description{
Verifică dacă o funcție reprezintă un pdf bivariat
}
