% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.6 e4.R
\name{media}
\alias{media}
\title{Calculează media unei funcții de densitate de probabilitate continue.}
\usage{
media(f)
}
\arguments{
\item{f}{funcția de densitate de probabilitate}
}
\value{
valoarea mediei
}
\description{
Calculează media unei funcții de densitate de probabilitate continue.
}
\examples{
media(function(x) { x^2 })
}
