% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.7 e6.R
\name{|.cRVrez}
\alias{|.cRVrez}
\title{OR Operator for cRVrez Objects}
\usage{
\method{|}{cRVrez}(Xres, Yres)
}
\arguments{
\item{Xres}{cRVrez object}

\item{Yres}{cRVrez object}
}
\value{
cRVrez object
}
\description{
OR Operator for cRVrez Objects
}
