% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3 e3.R
\name{plot_functie}
\alias{plot_functie}
\title{Plotează funcția de densitate (pdf) sau funcția de distribuție cumulativă (cdf) pentru o variabilă aleatoare unidimensională sau bidimensională.}
\arguments{
\item{fun}{Poate fi o funcție care primește un argument (x) sau două argumente (x, y). Poate fi un vector de valori y pentru distribuții unidimensionale sau un vector/matrice de valori z pentru distribuții bidimensionale.}

\item{xDomain}{Domeniul valorilor x pentru care se evaluează pdf/cdf.}

\item{yDomain}{Domeniul valorilor y pentru care se evaluează pdf/cdf. Dacă este gol sau nu este furnizat, 'plot_functie()' presupune o distribuție unidimensională.}
}
\description{
Plotează funcția de densitate (pdf) sau funcția de distribuție cumulativă (cdf) pentru o variabilă aleatoare unidimensională sau bidimensională.
}
\examples{

pdf2 <- function(x, y) { (0 <= x & x <= 1 & 0 <= y & y <= 1) * (2 * (1-x)) }

yDomain <- seq(0, 1, 0.1)
xDomain <- seq(0, 1, 0.1)

plot_functie(pdf2, xDomain, yDomain)
}
