% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.7 e6.R
\name{Pr}
\alias{Pr}
\title{Probability Calculation for cRVrez Objects}
\usage{
Pr(crez)
}
\arguments{
\item{crez}{cRVrez object}
}
\value{
numeric probability
}
\description{
Probability Calculation for cRVrez Objects
}
