% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.5 e8.R
\name{X_dist_marg}
\alias{X_dist_marg}
\title{Distribuția marginală pentru X dintr-un pdf bivariat}
\usage{
X_dist_marg(pdf)
}
\arguments{
\item{pdf}{trebuie să fie o funcție de densitate de probabilitate pentru o variabilă aleatoare bivariată}
}
\description{
Distribuția marginală pentru X dintr-un pdf bivariat
}
\examples{
pdf2 <- function(x, y) { (0 <= x & x <= 1 & 0 <= y & y <= 1) * (2/3 * (x + 2*y)) }
f_X <- X_dist_marg(pdf2)
f_X(0.4)
}
